refactor(naming): Renamed project to jack_compiler
Also updated the README. Both indicating that we've moved from just the
analyzer to the full compiler now.
