--- jack_analyzer/exceptions.py
@@ -1,7 +1,0 @@
-"""
-Custom Exceptions
-"""
-
-
-class InvalidOperationException(Exception):
-    """Operation is invalid"""
