+++ jack_compiler/exceptions.py
@@ -1,0 +1,7 @@
+"""
+Custom Exceptions
+"""
+
+
+class InvalidOperationException(Exception):
+    """Operation is invalid"""
